examples/visualization/requirements.txt
numpy
tqdm
seaborn
matplotlib
umap-learn